+++ lib/identity_validator.rb
@@ -2,17 +2,19 @@
 
 class IdentityValidator < ActiveModel::EachValidator
   def validate_each(record, attribute, value)
-    identity = ValidatesIdentity::Identity.new(record, value, options)
+    identity = ValidatesIdentity::Identity.new(record, attribute, value, options)
 
-    return if identity.valid?
-
-    ruby_prior_version_three =
-      Gem::Version.new(RUBY_VERSION) < Gem::Version.new('3.0.0')
-
-    if ruby_prior_version_three
-      record.errors.add(attribute, :invalid, options)
+    if identity.valid?
+      record.send("#{attribute}=", identity.formatted) if options[:format]
     else
-      record.errors.add(attribute, :invalid, **options)
+      ruby_prior_version_three =
+        Gem::Version.new(RUBY_VERSION) < Gem::Version.new('3.0.0')
+
+      if ruby_prior_version_three
+        record.errors.add(attribute, :invalid, options)
+      else
+        record.errors.add(attribute, :invalid, **options)
+      end
     end
   end
 end