+++ spec/validates_identity/identity_spec.rb
@@ -6,27 +6,27 @@ RSpec.describe ValidatesIdentity::Identity do
   let(:user) { User.new }
 
   it 'accepts nil value' do
-    expect(described_class.new(user, nil, {})).to be_valid
+    expect(described_class.new(user, :identity, nil, {})).to be_valid
   end
 
   it 'accepts blank value' do
-    expect(described_class.new(user, '', {})).to be_valid
+    expect(described_class.new(user, :identity, '', {})).to be_valid
   end
 
   it 'rejects if identity type is not defined' do
-    expect(described_class.new(user, '11144477735', {})).not_to be_valid
+    expect(described_class.new(user, :identity, '11144477735', {})).not_to be_valid
   end
 
   it 'rejects if identity type is nil' do
-    expect(described_class.new(user, '11144477735', { identity_type: nil })).not_to be_valid
+    expect(described_class.new(user, :identity, '11144477735', { identity_type: nil })).not_to be_valid
   end
 
   it 'rejects if identity type is blank' do
-    expect(described_class.new(user, '11144477735', { identity_type: '' })).not_to be_valid
+    expect(described_class.new(user, :identity, '11144477735', { identity_type: '' })).not_to be_valid
   end
 
   it 'rejects if identity type is not registered' do
-    expect(described_class.new(user, '11144477735', { identity_type: :abc_validator })).not_to be_valid
+    expect(described_class.new(user, :identity, '11144477735', { identity_type: :abc_validator })).not_to be_valid
   end
 
   context 'with a validator registered' do
@@ -36,7 +36,7 @@ RSpec.describe ValidatesIdentity::Identity do
     end
 
     it 'uses the validator defined' do
-      expect(described_class.new(user, '11144477735', identity_type: :identity_type)).to be_valid
+      expect(described_class.new(user, :identity, '11144477735', identity_type: :identity_type)).to be_valid
     end
   end
 
@@ -48,7 +48,7 @@ RSpec.describe ValidatesIdentity::Identity do
     end
 
     it 'uses the validator defined' do
-      expect(described_class.new(user, '11144477735', identity_type: :identity_type)).to be_valid
+      expect(described_class.new(user, :identity, '11144477735', identity_type: :identity_type)).to be_valid
     end
   end
 end